Refactor
빠진 필드 추가
라인정리
