+++ src/main/java/com/service/runnersmap/entity/ChatRoom.java
@@ -28,9 +28,8 @@ public class ChatRoom {
 //  @JoinColumn(name = "admin_id")
 //  private User admin; // 그룹장
 
-  @OneToOne
-  @JoinColumn(name = "post_id")
-  private Post post; // 연결된 모집글
-
+//  @OneToOne
+//  @JoinColumn(name = "post_id")
+//  private Post post; // 연결된 모집글
 
 }
