+++ src/main/java/com/service/runnersmap/service/ChatService.java
@@ -1,4 +1,4 @@
-package com.service.runnersmap;
+package com.service.runnersmap.service;
 
 import com.service.runnersmap.dto.ChatMessageDto;
 import com.service.runnersmap.entity.ChatMessage;
@@ -29,7 +29,8 @@ public class ChatService {
     chatMessageDto = ChatMessageDto.builder()
         .chatRoomId(chatMessageDto.getChatRoomId())
         .senderId(chatMessageDto.getSenderId())
-        .message(chatMessageDto.getSenderId() + "님이 채팅방에 입장하셨습니다.")
+        .senderNickname(chatMessageDto.getSenderNickname())
+        .message(chatMessageDto.getSenderNickname() + "님이 채팅방에 입장하셨습니다.")
         .build();
 
     // 브로커로 메시지 전송
@@ -37,13 +38,14 @@ public class ChatService {
   }
 
   // 메시지 저장
-  public void saveMessage(ChatMessageDto chatMessageDTO) {
+  public void saveMessage(ChatMessageDto chatMessageDto) {
 
-    // 채팅방 조회
-    ChatRoom chatRoom = chatRoomRepository.findById(chatMessageDTO.getChatRoomId())
+    ChatRoom chatRoom = chatRoomRepository.findById(chatMessageDto.getChatRoomId())
         .orElseThrow(() -> new RuntimeException("존재하지 않는 채팅방입니다."));
 
-    User sender = userRepository.findById(chatMessageDTO.getSenderId())
+
+    // 추후 jwt에서 사용자 정보 추출하는 로직 추가해야 함
+    User sender = userRepository.findById(chatMessageDto.getSenderId())
         .orElseThrow(() -> new RuntimeException("존재하지 않는 사용자입니다."));
 
 
@@ -52,14 +54,13 @@ public class ChatService {
         // 로그인 한 상태면 작성하지 않아도 될까요?
         // STOMP의 경우 헤드에 인증 관련한 걸 넣을 수 있다고 해서 좀 더 알아보겠습니다.
         .sender(sender)
-        .message(chatMessageDTO.getMessage())
+        .message(chatMessageDto.getMessage())
         .sentAt(LocalDateTime.now())
         .build();
 
-    // 메시지 저장
     chatMessageRepository.save(message);
 
-    template.convertAndSend("/sub/chat/room/" + chatMessageDTO.getChatRoomId(), chatMessageDTO);
+    template.convertAndSend("/sub/chat/room/" + chatMessageDto.getChatRoomId(), chatMessageDto);
 
   }
 
@@ -73,6 +74,7 @@ public class ChatService {
         .map(message -> ChatMessageDto.builder()
             .chatRoomId(message.getChatRoom().getId())
             .senderId(message.getSender().getId())
+            .senderNickname(message.getSender().getNickname())
             .message(message.getMessage())
             .sentAt(message.getSentAt())
             .build())
